Add the start of an neovim config
I want to start using neovim a lot more. One of the places I will end
up using it the most is in school, for notes. The best resource I've
seen for this is https://castel.dev/post/lecture-notes-1/. This commit
is the first step towards that.

Requirements.txt is just a little thing for me in the future. When I
get a new laptop or something and need to reinstall things, it'll be
nice to know what I need.

install.sh updated accordingly.